Add doWatchMessage.js
